feat(seo): add functionality to generate a `sitemap.xml` file (#788)
## Description

This PR adds some dependencies (and their configurations) to enable
building a
[sitemap](https://developers.google.com/search/docs/crawling-indexing/sitemaps/overview)
for the site. The file will help improve the SEO standings of the site
(hopefully). More details of the PR are as follows:

1. Installed the `nuxt-site-config` and `nuxtjs/sitemap` module to
generate the sitemap (with good DX).
2. Added the necessary configurations required to generate and
manipulate the sitemap according to the specific requirements of the
project.