chore: add support for Cloudflare Web Analytics (#789)
## Description

This PR adds support for collection analytics from the website using
Cloudflare Web Analytics. Following are the updates introduced with the
PR:

1. Added the necessary dependencies to load the script for Cloudflare to
start collecting the data.
2. Configure the Nuxt.js configurations to inject the token for
Cloudflare to be able to collect and showcase the data analytics.

**NOTE**: Although the PR adds support for the collecting data using
Cloudflare, whether it works or not will require a bit of patient
observation.