chore: add custom `ProsePre` component (#796)
## Description

This PR adds a custom `ProsePre` component which is partially inspired
from the `ProsePre` component provided in the
[nuxt-content/mdc](https://github.com/nuxt-content/mdc/blob/main/src/runtime/components/prose/ProsePre.vue)
package. The component is work-in-progress since it has not been
implemented to its expected behaviour. Regardless, the current
implementation will be a starting point to improve the component
further.

Following are some future improvements to make to the component:

1. Add a "copy" button which shows on hover.
2. Add number lines on the left without losing the syntax highlighting
of the code.
3. Provide custom logic to intelligently render a "shell" or a "console"
code fence differently.
4. Add a logic to handle diff'ing source code.

### Screenshots / Gifs

<img width="1920" height="1020" alt="image"
src="https://github.com/user-attachments/assets/bb02da2b-e6ac-4bb5-9757-13b54de58e6e"
/>